pyproject.toml
[tool.poetry]
name = "contexta"
version = "0.1.0"
description = "RAG application"
authors = ["Douglas Henrique <douglas.hsp2019@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.128.0"
uvicorn = "^0.40.0"
django = "^6.0.1"
python-dotenv = "^1.2.1"
pypdf = "^5.0.0"
openai = "^1.0.0"
qdrant-client = "^1.7.0"
djangorestframework = "^3.15.0"
httpx = "^0.27.0"
python-docx = "^1.1.0"
pytest = "^9.0.0"
pytest-cov = "^7.0.0"
pytest-mock = "^3.15.0"
pytest-asyncio = "^1.0.0"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "autopep8 (>=2.3.2,<3.0.0)",
    "isort (>=7.0.0,<8.0.0)",
    "black (>=25.12.0,<26.0.0)",
    "flake8 (>=7.3.0,<8.0.0)"
]
